% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reports.R
\name{get_reports}
\alias{get_reports}
\title{get_reports}
\usage{
get_reports(file_extension)
}
\arguments{
\item{file_extension}{the file extension that the text reports are stored in (the files must be in /data project subdirectory)}
}
\value{
the reports in a data frame
}
\description{
get_reports
}
