% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_regular_expressions.R
\name{get_variable_regular_expressions}
\alias{get_variable_regular_expressions}
\title{get_variable_regular_expressions}
\usage{
get_variable_regular_expressions()
}
\value{
returns a list of vectors, each vector contains regular expressions that will find a numerical value as their first capture group +/- the units as the second capture group
}
\description{
get_variable_regular_expressions
}
\examples{
get_variable_regular_expressions()

}
